feat: Finalize service-oriented architecture with dependency injection
