--- golem_workers/entrypoints/web/application_new.py
@@ -1,101 +1,0 @@
-import logging.config
-from contextlib import asynccontextmanager
-
-from fastapi import FastAPI, Request, status
-from fastapi.middleware.cors import CORSMiddleware
-from starlette.responses import JSONResponse
-
-from golem.utils.logging import DEFAULT_LOGGING
-from golem_workers import __version__
-from golem_workers.containers_new import Container
-from golem_workers.entrypoints.web.endpoints_new import router, Tags
-from golem_workers.exceptions import GolemWorkersError, ObjectNotFound, ObjectAlreadyExists
-from golem_workers.settings import Settings
-
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    container = app.state.container
-
-    await container.init_resources()
-    yield
-    await container.shutdown_resources()
-
-
-async def golem_workers_error_handler(request: Request, exc: GolemWorkersError):
-    return JSONResponse(
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-        content={"detail": f"Unhandled server error! {exc}"},
-    )
-
-
-async def object_not_found_handler(request: Request, exc: GolemWorkersError):
-    return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"detail": str(exc)})
-
-
-async def object_already_exists_handler(request: Request, exc: GolemWorkersError):
-    return JSONResponse(status_code=status.HTTP_409_CONFLICT, content={"detail": str(exc)})
-
-
-def create_application() -> FastAPI:
-    # Create and configure the container
-    container = Container()
-    container.settings.from_dict(dict(Settings()))
-    settings = container.settings()
-
-    # Configure logging
-    logging.config.dictConfig(DEFAULT_LOGGING)
-    logging.config.dictConfig(settings["logging_config"])
-
-    # Create the FastAPI application
-    app = FastAPI(
-        title="Golem Workers Specification",
-        version=__version__,
-        lifespan=lifespan,
-        exception_handlers={
-            GolemWorkersError: golem_workers_error_handler,
-            ObjectNotFound: object_not_found_handler,
-            ObjectAlreadyExists: object_already_exists_handler,
-        },
-        openapi_tags=[
-            {
-                "name": Tags.CLUSTERS,
-                "description": "Endpoints related to Cluster management.",
-            },
-            {
-                "name": Tags.NODES,
-                "description": "Endpoints related to Node management.",
-            },
-            {
-                "name": Tags.MISC,
-                "description": "General endpoints for utilities.",
-            },
-            {
-                "name": Tags.PORTS,
-                "description": "Endpoints for port allocation management.",
-            },
-        ],
-    )
-    
-    # Include the router
-    app.include_router(router)
-    
-    # Store the container in the app state
-    app.state.container = container
-
-    # Configure CORS
-    origins = [
-        "https://docs.golem.network",
-        "http://localhost:3000",
-    ]
-    vercel_origin_regex = "https://.*\.vercel\.app"
-    app.add_middleware(
-        CORSMiddleware,
-        allow_origins=origins,
-        allow_origin_regex=vercel_origin_regex,
-        allow_credentials=True,
-        allow_methods=["*"],
-        allow_headers=["*"],
-    )
-
-    return app
