+++ golem_workers/entrypoints/web/main.py
@@ -1,3 +1,3 @@
-from golem_workers.entrypoints.web import create_application
+from golem_workers.entrypoints.web.application_new import create_application
 
 app = create_application()
