--- golem_workers/entrypoints/web/main_new.py
@@ -1,3 +1,0 @@
-from golem_workers.entrypoints.web.application_new import create_application
-
-app = create_application()
