+++ golem_workers/services/__init__.py
@@ -1,4 +1,6 @@
 # Import services here for easy access
+__all__ = ["ProposalService", "ClusterService", "NodeService", "PortAllocationService"]
+
 from golem_workers.services.proposals import ProposalService
 from golem_workers.services.clusters import ClusterService
 from golem_workers.services.nodes import NodeService
