+++ golem_workers/services/clusters.py
@@ -1,5 +1,5 @@
 import asyncio
-from typing import Dict, List, MutableMapping, Any, Optional, Mapping
+from typing import Dict, List, MutableMapping, Any
 
 from golem.node import GolemNode
 
@@ -7,18 +7,13 @@ from golem_workers.cluster import Cluster
 from golem_workers.exceptions import ObjectAlreadyExists, ObjectNotFound
 from golem_workers.models import (
     ClusterOut,
-    AllocationConfig,
-    PaymentConfig,
-    NetworkConfig,
-    BudgetConfig,
-    NodeConfig,
 )
 from golem_workers.services.interfaces import IClusterService
 
 
 class ClusterService(IClusterService):
     """Service for managing clusters."""
-    
+
     def __init__(
         self,
         golem_node: GolemNode,
@@ -28,15 +23,15 @@ class ClusterService(IClusterService):
         self._golem_node = golem_node
         self._clusters_lock = clusters_lock
         self._clusters = clusters
-    
+
     async def create_cluster(self, request_data) -> ClusterOut:
         """Create a new cluster."""
         async with self._clusters_lock:
             cluster_id = request_data.cluster_id
-            
+
             if cluster_id in self._clusters:
                 raise ObjectAlreadyExists(f"Cluster with id `{cluster_id}` already exists!")
-            
+
             # Create cluster
             self._clusters[cluster_id] = cluster = Cluster(
                 golem_node=self._golem_node,
@@ -48,35 +43,35 @@ class ClusterService(IClusterService):
                 node_types=request_data.node_types,
                 labels=request_data.labels,
             )
-            
+
             cluster.schedule_start()
-            
+
             return ClusterOut.from_cluster(cluster)
-    
+
     async def list_clusters(self) -> List[str]:
         """List all available clusters."""
         async with self._clusters_lock:
             return list(self._clusters.keys())
-    
+
     async def get_cluster(self, cluster_id: str) -> ClusterOut:
         """Get details for a specific cluster."""
         cluster = self._clusters.get(cluster_id)
-        
+
         if not cluster:
             raise ObjectNotFound(f"Cluster with id `{cluster_id}` does not exists!")
-        
+
         return ClusterOut.from_cluster(cluster)
-    
+
     async def delete_cluster(self, cluster_id: str) -> Dict[str, Any]:
         """Delete a cluster."""
         async with self._clusters_lock:
             cluster = self._clusters.get(cluster_id)
-            
+
             if not cluster:
                 raise ObjectNotFound(f"Cluster with id `{cluster_id}` does not exists!")
-            
+
             # Stop and remove cluster
             await cluster.stop()
             del self._clusters[cluster_id]
-            
+
             return {"cluster": ClusterOut.from_cluster(cluster)}
