+++ golem_workers/work.py
@@ -8,7 +8,6 @@ from pathlib import Path
 
 from golem.resources import BatchError, DeployArgsType, Network
 
-from golem_workers.cluster import Node
 from golem_workers.context import WorkContext
 from golem_workers.utils import (
     create_ssh_key,
@@ -124,8 +123,17 @@ async def run_in_shell(context: WorkContext, /, *commands: Union[str, Sequence[s
 
         raise
 
-async def wait_for_http_endpoint(context : WorkContext, *, network_name : str, port : int, path : str = '/', timeout_minutes: float = 15) -> None:
+
+async def wait_for_http_endpoint(
+    context: WorkContext,
+    *,
+    network_name: str,
+    port: int,
+    path: str = "/",
+    timeout_minutes: float = 15,
+) -> None:
     import websockets
+
     key = context.golem_node.app_key
     logger.info("Waiting for http endpoint %s:%s%s", network_name, port, path)
     node = context.extra["node"]
@@ -135,13 +143,16 @@ async def wait_for_http_endpoint(context : WorkContext, *, network_name : str, p
 
     async def try_connect():
         try:
-            async with websockets.connect(f"{uri}/{port}", additional_headers={
-                'Authorization': f'Bearer {key}',
-            }) as websocket:
+            async with websockets.connect(
+                f"{uri}/{port}",
+                additional_headers={
+                    "Authorization": f"Bearer {key}",
+                },
+            ) as websocket:
                 await websocket.send(f"GET {path} HTTP/1.1\r\nHost: {node_ip}:{port}\r\n\r\n")
                 data = await websocket.recv(False)
                 return data[0:12] == b"HTTP/1.1 200"
-        except:
+        except Exception:  # noqa: BLE001
             return False
 
     async def wait_for_endpoint():
@@ -151,6 +162,7 @@ async def wait_for_http_endpoint(context : WorkContext, *, network_name : str, p
     timeout_seconds = timedelta(minutes=timeout_minutes).total_seconds()
     await asyncio.wait_for(wait_for_endpoint(), timeout=timeout_seconds)
     logger.info("Http endpoint %s:%s%s is ready", network_name, port, path)
+
 
 async def stop_activity(context: WorkContext, /) -> None:
     """Stop and destroy activity.
